Major: 0.0.118
New Features:
1. Import Black-Bullet-backend database to read/write model.
2. Initial support runtime read/write database.

+++ src/Modules/Music/Libraries/knmusicinfocollector.cpp
@@ -205,7 +205,7 @@ void KNMusicInfoCollector::parseByMediaInfo(const QString &value)
     numberData=rawInfoData.left(secondPos);
     numberData.remove(' ');
     setMediaData(KNMusicGlobal::SampleRate, numberData+" "+rawInfoData.mid(secondPos+1));
-    m_samplingRate=numberData.toInt();
+    m_samplingRate=numberData.toFloat();
 }
 
 void KNMusicInfoCollector::readID3v1Tag(const QString &value)
